+++ kitty/options/parse.py
@@ -12,9 +12,9 @@ from kitty.options.utils import (
     deprecated_hide_window_decorations_aliases, deprecated_macos_show_window_title_in_menubar_alias,
     deprecated_send_text, disable_ligatures, edge_width, env, font_features, hide_window_decorations,
     macos_option_as_alt, macos_titlebar_color, narrow_symbols, optional_edge_width, parse_map,
-    parse_mouse_map, resize_draw_strategy, scrollback_lines, scrollback_pager_history_size,
-    shell_integration, store_multiple, symbol_map, tab_activity_symbol, tab_bar_edge,
-    tab_bar_margin_height, tab_bar_min_tabs, tab_fade, tab_font_style, tab_separator,
+    parse_mouse_map, paste_actions, resize_draw_strategy, scrollback_lines,
+    scrollback_pager_history_size, shell_integration, store_multiple, symbol_map, tab_activity_symbol,
+    tab_bar_edge, tab_bar_margin_height, tab_bar_min_tabs, tab_fade, tab_font_style, tab_separator,
     tab_title_template, titlebar_color, to_cursor_shape, to_font_size, to_layout_names, to_modifiers,
     url_prefixes, url_style, visual_window_select_characters, window_border_width, window_size
 )
@@ -1085,6 +1085,9 @@ class Parser:
 
     def open_url_with(self, val: str, ans: typing.Dict[str, typing.Any]) -> None:
         ans['open_url_with'] = to_cmdline(val)
+
+    def paste_actions(self, val: str, ans: typing.Dict[str, typing.Any]) -> None:
+        ans['paste_actions'] = paste_actions(val)
 
     def placement_strategy(self, val: str, ans: typing.Dict[str, typing.Any]) -> None:
         val = val.lower()